Close the handle returned by WrapNativeCodec2GrallocHandle()
WrapNativeCodec2GrallocHandle() creates a new native handle, and
clones the fd of the input native handle. The caller should explicitly
close and delete the returned native handle.

This CL updates the description of WrapNativeCodec2GrallocHandle(),
and calls native_handle_close() and native_handle_delete() if the
ownership of the returned handle is not transfered to
C2GraphicAllocation successfully.

Bug: 181541291
Test: android.media.cts.AdaptivePlaybackTest

Change-Id: I53726b61115a212675a199ddcd843df1fead948c