Camera: Fix handling of dead processes for arbitration
A dead (or dying) process will have state NONEXISTENT, OOM score
INVALID when queried from ProcessInfoService.  Ensure that this maps
to a very low total score for prioritization.

Also add unit testing for this, and improve the unit tests and service
code to use named constants instead of magic numbers for state/score.

Test: New unit tests pass; CTS not regressed on coral
Bug: 181777896
Change-Id: I19f1a7e0b75f35e64ff790ef7098b3d40b7e55be