WriterFuzzer: Fix multiple library linkage
In libstagefright_writer_fuzzer "libstagefright_webm" library are
getting linked multiple times which leads to multiple destruction.
Multiple destruction of the same library lead to a crash.

libstagefright_webm, libstagefright_esds, libogg libraries are
statically linked in "libstagefright" which is dynamically linked
in "libstagefright_fuzzer_defaults" which is included while defining
"libstagefright_writer_fuzzer". So static linking of above three
libraries again in libstagefright_writer_fuzzer is not required
because it will lead to multiple linkages and multiple destructions.

Bug: 180629956
Test: test in bug

Change-Id: I789d900d52b7f4a2171ace45128c3ecf115a9213