mediacodec: hoist NULL assignments to variable initialization
There're paths through this function where we end up using `key` (and
presumably `iv`) without setting them. These paths _might_ be
dynamically unreachable (I'm unfamiliar with this codebase), but it's
locally unclear that that's the case.

As a security mitigation, our compiler will zero-initialize these
in all cases anyway. That said, it's probably best to make it more
obvious to the reader that these are intended to be NULL on all paths
except the `else if (!c2Buffer)` one

Bug: None
Test: TreeHugger
Change-Id: Ib3c8e7d6b7cc8b794118c8784419458191d0cb56