audiopolicy: reflect libbinder/am split
Splitting out extra classes from libbinder so that it takes less memory
to use (many processes don't use activity manager). Currently, this
library is empty, and this CL is being merged to add the dependency
before activity manager code is moved into this library, because this
file was only converted to Android.bp on internal branches.

Bug: 183654927
Test: N/A
Change-Id: Idd20f64a2bb54c1364aa996d6e4ec4fe9e27d732