audio policy: add back use of A2DP for phone strategy
Refactoring to use preferred device for srategy instead of forced
usage introduced a regression where A2DP device is not used any more for
phone strategy when not in a call.

Fixed this by using the same list of removable devices for phone
strategy as for media strategy. The preferred device will still be
applied before this default list to honor SCO or speaker phone modes.

Bug: 184771057
Test: Play VOICE_COMMUNICATION audio with A2DP only headphones
Change-Id: I032d962d6ebc311b055bb35fb34d5310fc308bc7