audio policy: fix call audio switching from BT SCO to earpiece
commit c0697594 made that A2DP is unconditionally used for phone
strategy when connected. This creates a problem when disabling SCO
as the call is now routed to A2DP instead of earpiece by default which
changes from the expected route and does not work if the system suspends
A2DP audio when a call starts.

Restore previous policy by removing A2DP devices from eligible
devices for phone strategy when in call. This preserves the fix for
b/184771057 by still routing voice mails to A2DP when connected.

Bug: 188020987
Bug: 184771057

Test: manual call audio regression.
Change-Id: Ibab19c6150b84f8e8a46e4389b66f20cff205c10