AudioFlinger: fix player ID generation after crash
The uniqued ID generator will generate the same player  IDs after a native
audioserver restarts as it always resumes genrating from 1.
This can be a problem as the player ID of existing clients is
preserved by java audioservice and reused when the client reconnects after a
native audioserver process crash.

This change is an extension to commit f43be0c7 to cover Player IDs in addition
to session Ids.

Bug: 189810552
Test: make
Change-Id: Ic4796958bb520f9d13f97ae27b3776506f70082f