media_codecs_sw: Update supported resolution for av1
Updated av1 decoder's supported resolution to 2048x2048.
Problem with 1920x1080 is, with a block size that is 16x16,
MediaCodecInfo considers 1920x1072 as supported range (floors
the dimensions to multiple of 16).
Also it doesn't support portrait videos where height is greater
than 1072.
Now it is updated to be consistent with what plugin supports.

Bug: 167276410
Test: app in the bug
Change-Id: Idc8d3bdbd7f126b4e70c265635e87eaf4208d848