Avoid MTP open binder call to MediaProvider
For transcoding, we forced all MTP file opens on external storage to
use the ContentResolver#openTypedAssetFile. This allowed us to
flexibly support transcoding Option A/Option B where we might want to
transcode by default or not transcode by default. Unfortunately, this
caused performance issues when transferring a lot of files via MTP.

Since in Android S, we are launching with Option A, which doesn't
transcode by default we can limit the ContentResolver opens to only
cases where transcode=true. Transcode=true is really an advanced user
option and should be false for most cases.

Test: Manual
Bug: 189977601
Change-Id: I9b2f94189e69a8f9d11803527a0095537f049c84