Camera: Avoid possible race conditions during in-flight dump
Access to the camera device "mInFlightMap" is currently
synchronized via "mInFlightLock". The device dump
sequence doesn't try to lock or acquire this mutex when
working with the map which makes data races  possible.
To avoid potential instabilities and inconsistent dump
information, try to lock 'mInFlightLock' before querying
any in-flight data.

Bug: 191224953
Test: Manual using camera application
Change-Id: Icb5c172a17476c7ef22bb85ca9430dfbf014e947