Dolbyvision parsing used disallowed functions
The Dolbyvision box parsing uses raw access to the AMessage routines
underneath the AMediaFormat_* routines. This violates the mainline
api boundary rules and causes crashes now that the underlying
AMessage implementation has changed.

The code wanted to delete a CSD stored in an AMediaFormat, but there is
no ndk/stable interface to do that.  The available solution was to
replace that buffer with an empty one; I don't know if this will impact
the dolbyvision decoders.

Bug: 129367182
Bug: 193017338
Test: atest android.media.cts.MediaExtractor
Change-Id: Ib70db91c7b5209a95b8bcfd89fbb2dfc41194eba