audio policy: fix record op monitor use of wrong command thread
commit ed726ccd introduced a potential deadlock in the audio policy
commnand thread dedicated to commands going from audio policy manager
to audioflinger. The thread was used by mistake when posting a command
to update the UID states which requires to take the audio policy service
lock. This type of commands should never be run in the audio command
thread but in the output command thread instead.

Bug: 193164403
Test: atest AudioPlaybackCaptureTest#testStopMediaProjectionDuringCapture
Change-Id: I150c768dd562f2658eaf4a31b837cf677136a3af