C2 VTS: Use string::compare instead of string::find to match mime
Instead of std::string::find() to match mime of components with the
entries in the table, use std::string::compare() to find an exact match.
Without this a codec can wrongly be matched with an entry in the table
if its mime contains one of the entries in the table (entry is substring
of component's mime).

Because we now match the complete mime, few entries in the table are
updated to exact values needed.

Bug: 194774947
Test: atest VtsHalMediaC2V1_0TargetAudioDecTest
Test: atest VtsHalMediaC2V1_0TargetVideoDecTest

Change-Id: I4ee38590af66cb31739be0bdfd60075d77b61c70