Add format conversion in AudioRecord.
In audio flinger, it is required that the client's format matches the
hardware format to get fast mode. However, format conversion is not a
time consuming operation. In that case, adding format conversion in
AudioRecord can help the client get fast mode if only format is
different from the hardware one. This can help improve latency.

Test: atest AudioRecordTest, capture audio in fast/non-fast mode
Bug: 79156275
Bug: 188843084
Change-Id: I1baaa9a6054bb84412c31a159c6d25ee498822cf