AudioFlinger: Workaround for dumpsys stall
Binder fds may not close immediately due to a race
condition when the last refcount is removed on the remote
binder thread and it is blocked in the kernel waiting for data.
Issue a binder ping command to unblock the remote binder thread
and complete the fd close / release.

Test: adb shell dumpsys media.audio_flinger (repeat many times)
Bug: 111997867
Bug: 177271958
Change-Id: I2237d1af430dfad08f22de6efaec9b04c19bdc5a
(cherry picked from commit e72ff02ba1c862968619ed0dead2fbd8b41dec09)