Better support for head and screen sensors being the same one
This is not a super interesting use-case, but for completeness we
should make sure it is handled correctly.

From the Sensor Framework standpoint, a sensor can only be enabled
once at a given time. We mirror this behavior in SensorPoseProvider
and document it well.

At the SpatializerPoseController layer we make sure that we don't
enable a sensor more than once and that if both sensors are the same
that we treat an event as applying to both.

Test: manual verification of this mode.
Bug: 188502620
Change-Id: I3aa1f225d7100c5d934af9255484caec75375380