Split AudioProfile between fwk and android.media.audio.common
a.m.a.c adds a version of AudioProfile struct which
is similar to Audio HIDL HAL V7. It lacks some bits used only
by the framework. These are carried in AudioProfileSys.
AudioProfileSys does not contain AudioProfile. This is done
to simplify further splitting of other types like AudioPort.

Updated to/from AIDL conversions in libaudiofondation.
They encapsulate the fact that there are two AIDL structures
per AudioProfile class instance.

Bug: 198812639
Test: atest audiofoundation_parcelable_test
Change-Id: I6834650f9406c552b6dfa42b120769eb38d6f588