let outputFormat update and discard buffer
For tunneled playback, there should be no buffer output. But for Codec2,
CCodec still needs to update OutputFormat per frame to MediaCodec. So,
MediaCodec should discard output buffer and only update OutputFormat in
this case. And for Block Model, the null buffer still needs to pass to
application to provide OutputFormat per frame.

Test: Digital TV test pass

Bug: 193867146
Bug: 199668436
Change-Id: Ifbad1de8992cc212d62e21ca6a4db832ee7b2d76