Split AudioGain* between fwk and android.media.audio.common
AudioGainMode and AudioGainConfig are moved as is.
The version of AudioGain in a.m.a.c is similar to the one
from Audio HIDL HAL V7. It lacks some bits used only
by the framework. These are carried in AudioGainSys.
AudioGainSys does not contain AudioGain. This is done
to simplify further splitting of other types like AudioPort.

Updated to/from AIDL conversions in libaudiofondation.
They encapsulate the fact that there are two AIDL structures
per AudioGain class instance. The AudioGain class is
simplified by stripping off no more required implementation
of Parcelable, and reusing conversion functions from
AidlConversion.

Bug: 198812639
Test: atest audiofoundation_parcelable_test
Change-Id: Ic8fa34f6d873e026c785497bb48c4a12fa90a442