Remove automatic addition of AC3, IEC61937 to HDMI outputs
Historically, the automatic addition of AC3 and IEC61937 formats
to the list of supported formats for HDMI output ports was a
workaround for the problem occuring in certain TV setups that
don't report supported formats properly. It was found out that
this workaround creates confusion in setups which report
supported formats properly. This is why automatic addition is
being removed now.

It's still possible to restore the historic behavior if required:
 - on vendor side, by adding
  'encodedFormats="AUDIO_FORMAT_AC3 AUDIO_FORMAT_IEC61937"' to the
   HDMI output device port in the APM config file;
 - on user side, by going to "Surround Sound" settings and
   enabling "Dolby Digital" manually.

Bug: 199687164
Bug: 201089251
Test: atest audiopolicy_tests
Test: manual test on an Android TV dongle
Merged-In: I54d9a2c2f203db9fb462ef6d1206ce17043ab997
Change-Id: I54d9a2c2f203db9fb462ef6d1206ce17043ab997