Close `watch print` gracefully with user interrupt
The print functionality of watch command needed to be stopped with a
SIGINT. However, this could lead to some resources (for example the
output file descriptor) not being closed properly.

This CL adds a wrapper around the poll function allowing print command
to wait for the user to press the return key and quit gracefully on user
input.

Bug: 199746421
Test: Tested Manually
Change-Id: I2bb3ffd6b0711f40ecafe3a94c21e756ac2cd4e6