Invalidate compressed tracks when secondary outputs are changed.
As compressed tracks are not supported for playback capture, it will
cause crash when secondary outputs are changed and it is only updated
with secondary outputs. In that case, invalidating offload tracks
when secondary outputs are changed to trigger recreation.

Test: atest AudioPlaybackCaptureTest
Bug: 207152962
Change-Id: Iff6ea6f4797dbd160b1e0169f375062725a670af
Merged-In: Iff6ea6f4797dbd160b1e0169f375062725a670af